+++ dflib/src/main/java/org/dflib/exp/agg/BigintAggregators.java
@@ -1,14 +1,21 @@
 package org.dflib.exp.agg;
 
+import org.dflib.DecimalExp;
 import org.dflib.Series;
 import org.dflib.builder.ObjectAccum;
 
+import java.math.BigDecimal;
 import java.math.BigInteger;
+
+import static org.dflib.Exp.$bigint;
+import static org.dflib.Exp.count;
 
 /**
  * @since 2.0.0
  */
 public class BigintAggregators {
+
+    private static final DecimalExp avg = $bigint(0).sum().castAsDecimal().div(count());
 
     public static Series<BigInteger> cumSum(Series<BigInteger> s) {
 
@@ -59,5 +66,9 @@ public class BigintAggregators {
         }
 
         return sum;
+    }
+
+    public static BigDecimal avg(Series<BigInteger> s) {
+        return s.size() == 0 ? BigDecimal.ZERO : avg.reduce(s);
     }
 }